moved functions into userData.ts
